Extract AvatarsSection to separate file
